+++ STGenetics.Domain/Model/AnimalModel.cs
@@ -9,7 +9,7 @@ namespace STGenetics.Domain.Model
 
     public class AnimalModel
     {
-        public int AnimalId { get; set; }
+        public long AnimalId { get; set; }
         public string Name { get; set; } = default!;
         public string Breed { get; set; } = default!;
         public DateTime BirthDate { get; set; }
